fix: separate files
1 function 1 file
