+++ srcs/execute/pwd.c
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tkomatsu <tkomatsu@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/01/22 23:24:08 by tkomatsu          #+#    #+#             */
+/*   Updated: 2021/01/22 23:31:39 by tkomatsu         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	msh_pwd(void)
+{
+	char	*path;
+
+	path = getcwd(NULL, 0);
+	ft_putendl_fd(path, 1);
+	ft_free(path);
+	return (1);
+}
