+++ srcs/execute/unset.c
@@ -1,36 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtin.c                                          :+:      :+:    :+:   */
+/*   unset.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tkomatsu <tkomatsu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/01/18 15:05:15 by tkomatsu          #+#    #+#             */
-/*   Updated: 2021/01/21 22:27:08 by tkomatsu         ###   ########.fr       */
+/*   Created: 2021/01/22 23:26:45 by tkomatsu          #+#    #+#             */
+/*   Updated: 2021/01/22 23:27:18 by tkomatsu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	msh_exit(void)
+int	msh_unset(char **args)
 {
-	return (0);
-}
+	int	i;
 
-int	msh_pwd(void)
-{
-	char	*path;
-
-	path = getcwd(NULL, 0);
-	ft_putendl_fd(path, 1);
-	ft_free(path);
-	return (1);
-}
-
-int	msh_cd(char	**args)
-{
-	if (chdir(args[1]) < 0)
-		perror("cd");
-	ft_setenv("PWD", getcwd(NULL, 0), 1);
+	i = 1;
+	while (args[i])
+		ft_unsetenv(args[i++]);
 	return (1);
 }
